Remove BaseWith trait, not working with traits nesting
