+++ src/Command/Traits/With/WithHash.php
@@ -2,17 +2,33 @@
 
 namespace Predis\Command\Traits\With;
 
-class WithHash
+use UnexpectedValueException;
+
+trait WithHash
 {
-    use BaseWith;
-
-    private function getKeyword(): string
+    public function setArguments(array $arguments)
     {
-        return 'WITHHASH';
-    }
+        $argumentsLength = count($arguments);
 
-    private function getArgumentPositionOffset(): int
-    {
-        return static::$withHashArgumentPositionOffset;
+        if (
+            $arguments[static::$withHashArgumentPositionOffset] >= $argumentsLength
+            || false === $arguments[static::$withHashArgumentPositionOffset]
+        ) {
+            parent::setArguments($arguments);
+            return;
+        }
+
+        $argument = $arguments[static::$withHashArgumentPositionOffset];
+
+        if (true === $argument) {
+            $argument = 'WITHHASH';
+        } else {
+            throw new UnexpectedValueException("Wrong WITHHASH argument type");
+        }
+
+        $argumentsBefore = array_slice($arguments, 0, static::$withHashArgumentPositionOffset);
+        $argumentsAfter = array_slice($arguments,  static::$withHashArgumentPositionOffset + 1);
+
+        parent::setArguments(array_merge($argumentsBefore, [$argument], $argumentsAfter));
     }
 }
